Modifications made to HYDMOD
The existing version of HYDMOD only allowed interpolation if all four
points were in active cells. A couple modifications were made to the
HDS and DDN observation types (they were not also made for the
subsidence package observation types)
1. A new interpolation type 'H' was created in addition to 'C' and 'I'.
The existing types were not changed (save for an update to a confusing
error message for 'I' in the previous methods)
2. Dummy values were assigned if any of the four interpolation points
falls in a DRY, NHOFLO, or otherwise inactive cell (e.g. the target for
a hydrograph is at the edge of the grid).
3. Rather than assigning interpolation points outside the grid, edge
cells were assigned
4. Cells with IBOUND=0, or head =HNOFLOW or head=HDRY were given dummy
variables based on surrounding cells following the logic of
MOD2OBS/MOD2SMP in the PEST suite of tools.

A folder called hydmod_test contains an iPython notebook comparing the
new HYDMOD features with MOD2SMP.